+++ src/main/groovy/de/triplet/gradle/play/ImageFileFilter.groovy
@@ -3,7 +3,7 @@ package de.triplet.gradle.play
 class ImageFileFilter implements FileFilter {
     @Override
     boolean accept(File pathname) {
-        return pathname.name.toLowerCase().endsWith(".png") ||
-                pathname.name.toLowerCase().endsWith(".jpg")
+        return pathname.name.toLowerCase().endsWith('.png') ||
+                pathname.name.toLowerCase().endsWith('.jpg')
     }
 }