+++ src/main/groovy/de/triplet/gradle/play/PlayPublisherPlugin.groovy
@@ -1,21 +1,19 @@
 package de.triplet.gradle.play
 
 import com.android.build.gradle.AppPlugin
-import org.apache.commons.lang.StringUtils
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 
 class PlayPublisherPlugin implements Plugin<Project> {
 
-    public static final String PLAY_STORE_GROUP = "Play Store"
+    public static final PLAY_STORE_GROUP = 'Play Store'
 
     @Override
     void apply(Project project) {
         def log = project.logger
 
-        def hasAppPlugin = project.plugins.find { p -> p instanceof AppPlugin }
-        if (!hasAppPlugin) {
-            throw new IllegalStateException("The 'com.android.application' plugin is required.")
+        if (!project.plugins.any { p -> p instanceof AppPlugin }) {
+            throw new IllegalStateException('The \'com.android.application\' plugin is required.')
         }
 
         def extension = project.extensions.create('play', PlayPublisherPluginExtension)
@@ -34,51 +32,38 @@ class PlayPublisherPlugin implements Plugin<Project> {
                 return
             }
 
-            def buildTypeName = variant.buildType.name.capitalize()
-
-            def productFlavorNames = variant.productFlavors.collect { it.name.capitalize() }
-            if (productFlavorNames.isEmpty()) {
-                productFlavorNames = [""]
-            }
-            def productFlavorName = productFlavorNames.join('')
-            def flavor = StringUtils.uncapitalize(productFlavorName)
-
-            def variationName = "${productFlavorName}${buildTypeName}"
-
-            def bootstrapTaskName = "bootstrap${variationName}PlayResources"
-            def playResourcesTaskName = "generate${variationName}PlayResources"
-            def publishApkTaskName = "publishApk${variationName}"
-            def publishListingTaskName = "publishListing${variationName}"
-            def publishTaskName = "publish${variationName}"
+            def bootstrapTaskName = "bootstrap${variant.name.capitalize()}PlayResources"
+            def playResourcesTaskName = "generate${variant.name.capitalize()}PlayResources"
+            def publishApkTaskName = "publishApk${variant.name.capitalize()}"
+            def publishListingTaskName = "publishListing${variant.name.capitalize()}"
+            def publishTaskName = "publish${variant.name.capitalize()}"
 
             // Create and configure bootstrap task for this variant.
             def bootstrapTask = project.tasks.create(bootstrapTaskName, BootstrapTask)
             bootstrapTask.extension = extension
             bootstrapTask.variant = variant
-            if (StringUtils.isNotEmpty(flavor)) {
+            if (!variant.flavorName.isEmpty()) {
                 bootstrapTask.playAccountConfig = project.android.productFlavors[flavor].playAccountConfig
-                bootstrapTask.outputFolder = new File(project.projectDir, "src/${flavor}/play")
+                bootstrapTask.outputFolder = new File(project.projectDir, "src/${variant.flavorName}/play")
             } else {
                 bootstrapTask.playAccountConfig = project.android.defaultConfig.ext.playAccountConfig
-                bootstrapTask.outputFolder = new File(project.projectDir, "src/main/play")
+                bootstrapTask.outputFolder = new File(project.projectDir, 'src/main/play')
             }
-            bootstrapTask.description = "Downloads the play store listing for the ${variationName} build. No download of image resources. See #18."
+            bootstrapTask.description = "Downloads the play store listing for the ${variant.name.capitalize()} build. No download of image resources. See #18."
             bootstrapTask.group = PLAY_STORE_GROUP
 
             // Create and configure task to collect the play store resources.
             def playResourcesTask = project.tasks.create(playResourcesTaskName, GeneratePlayResourcesTask)
 
-            playResourcesTask.inputs.file(new File(project.projectDir, "src/main/play"))
-            if (StringUtils.isNotEmpty(flavor)) {
-                playResourcesTask.inputs.file(new File(project.projectDir, "src/${flavor}/play"))
+            playResourcesTask.inputs.file(new File(project.projectDir, 'src/main/play'))
+            if (!variant.flavorName.isEmpty()) {
+                playResourcesTask.inputs.file(new File(project.projectDir, "src/${variant.flavorName}/play"))
             }
             playResourcesTask.inputs.file(new File(project.projectDir, "src/${variant.buildType.name}/play"))
-            if (StringUtils.isNotEmpty(flavor)) {
-                playResourcesTask.inputs.file(new File(project.projectDir, "src/${variant.name}/play"))
-            }
+            playResourcesTask.inputs.file(new File(project.projectDir, "src/${variant.name}/play"))
 
             playResourcesTask.outputFolder = new File(project.projectDir, "build/outputs/play/${variant.name}")
-            playResourcesTask.description = "Collects play store resources for the ${variationName} build"
+            playResourcesTask.description = "Collects play store resources for the ${variant.name.capitalize()} build"
             playResourcesTask.group = PLAY_STORE_GROUP
 
             // Create and configure publisher meta task for this variant
@@ -91,7 +76,7 @@ class PlayPublisherPlugin implements Plugin<Project> {
             }
             publishListingTask.variant = variant
             publishListingTask.inputFolder = playResourcesTask.outputFolder
-            publishListingTask.description = "Updates the play store listing for the ${variationName} build"
+            publishListingTask.description = "Updates the play store listing for the ${variant.name.capitalize()} build"
             publishListingTask.group = PLAY_STORE_GROUP
 
             // Attach tasks to task graph.
@@ -108,11 +93,11 @@ class PlayPublisherPlugin implements Plugin<Project> {
                 }
                 publishApkTask.variant = variant
                 publishApkTask.inputFolder = playResourcesTask.outputFolder
-                publishApkTask.description = "Uploads the APK for the ${variationName} build"
+                publishApkTask.description = "Uploads the APK for the ${variant.name.capitalize()} build"
                 publishApkTask.group = PLAY_STORE_GROUP
 
                 def publishTask = project.tasks.create(publishTaskName)
-                publishTask.description = "Updates APK and play store listing for the ${variationName} build"
+                publishTask.description = "Updates APK and play store listing for the ${variant.name.capitalize()} build"
                 publishTask.group = PLAY_STORE_GROUP
 
                 // Attach tasks to task graph.
@@ -122,7 +107,7 @@ class PlayPublisherPlugin implements Plugin<Project> {
 
                 variant.outputs.each { output -> publishApkTask.dependsOn output.assemble }
             } else {
-                log.warn("Signing not ready. Did you specify a signingConfig for the variation ${variationName}?")
+                log.warn("Signing not ready. Did you specify a signingConfig for the variation ${variant.name.capitalize()}?")
             }
         }
     }